% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalia.R
\name{margin_content}
\alias{margin_content}
\title{Allow margin_note()-like content including inline chunks.}
\usage{
margin_content(text, move = 0)
}
\arguments{
\item{text}{Character string containing the text to be put in the margin. An inline chunk 
can be delimiter like this: @r 3 + 2@. That way, the function call (which must itself be in an
inline chunk) can itself contain an inline chunk for creating a table, etc.}

\item{move}{Distance in cm to move the note. positive means down, negative means up. Works only in PDF mode.}
}
\description{
Allow margin_note()-like content including inline chunks.
}
\examples{
\dontrun{In an Rmd file, margin_content("We can add 3 + 2 to get @r 3 + 2@")}
}

